requirements.txt
pywebview>=5.0
qtpy>=2.4
PyQt6>=6.5
PyQt6-WebEngine>=6.5
mss>=9.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Local vision model (Qwen3-VL-2B)
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
modelscope>=1.10.0
qwen-vl-utils>=0.0.8

# Local TTS (Piper)
piper-tts>=1.2.0
